backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==0.28.1
firebase-admin==6.2.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.0